etl/requirements.txt
tushare
psycopg2-binary
python-dotenv
pandas
jupyter
matplotlib
sqlalchemy